Added reset function to MessageBag
Couldn't find a way to get around the exception thrown by the buildXml() function when having multiple requests per page load (e.g. getting a UserID from Sentora then sending another request with the UserID as a parameter, or creating a user then creating a domain for that user). Hence the addition of the reset function.